requirements.txt
Django>=3.1.3,<3.2.0
djangorestframework>=3.12.2,<3.13.0
PyJWT==1.7.1